Correlation Matrix and Atmospheric Noise
Using as reference the John work, I tried to evaluate the greatness of
the atmospheric noise in the TOD of a typically CMB ground-based
experiments.

To reach the aim, for firstly I would to compute the
C_{tt'}^{ij} correlation matrix working with low constraint equations
(in term of the radiation beam pattern, and various atmospheric profiles
like water vapor and spatial distribution of turbulence).
Secondly, deduce from the C_{tt'}^{ij} correlation matrix a set of skies
that could produce these sign in the TOD and evaluate an average behavior
and their impact on the CMB power spectrum.

Nowadays nothing is coded. A very heuristic approach to C_{tt'}^{ij} matrix
is present in `Integration` directory, using Monte-Carlo integration to
perform all possible samples correlations but it is very early.